Feat/map maker refactor (#44)
* start of map maker refactor

* remove annoying eslint rules

* fix: build error

* build: and another

* support for pasting google links + UI changes

* create reusable type for google maps config

* wire up delete button

* hide pastCoverage for this release

* save modal + modal UI updates + generic breakpoint hook

* poc for deckgl

* got updating/deleting & selected marker working

* get google link paste working again

* bit of tidy up

* finished up saveMapModal functionality

* extract parseJsonFile into its own helper

* support JSON objects that contain a valid array

* increase allowed body limit and cap custom locations to 50k for now

* remove locations on delete again

* ensure correct allotment orientation on mount

* remove unused deck react

* move new map editor to actual page

* fix: header loading state

* remove unused styles

* remove logs

* fix: loading state for selection map

* fix: pano description overflow

* updated layout to a single top and bottom menu bar

* refactor to a single toast function that accepts variants

* Increase picking radius for map markers

* remove unused styles

* remove unused redux slice

* remove unused type

* fix: bug with hasMadeChanges + original map details not resetting after save

* set radius to default 50m & adjust location + swap preview buttons

* fix up global button for release

* remove local map css

* revert preview buttons orientation

+++ backend/routes/maps/updateCustomMap.ts
@@ -2,6 +2,8 @@ import { ObjectId } from 'mongodb'
 import { NextApiRequest, NextApiResponse } from 'next'
 import { calculateMapScoreFactor, collections, getUserId, throwError } from '@backend/utils'
 import { LocationType } from '@types'
+import { MAX_ALLOWED_CUSTOM_LOCATIONS } from '@utils/constants/random'
+import { formatLargeNumber } from '@utils/helpers'
 
 type ReqBody = {
   name?: string
@@ -11,7 +13,13 @@ type ReqBody = {
   locations?: LocationType[]
 }
 
-type UpdatedMap = { name?: string; description?: string; previewImg?: string; isPublished?: boolean }
+type UpdatedMap = {
+  name?: string
+  description?: string
+  previewImg?: string
+  isPublished?: boolean
+  lastUpdatedAt?: Date
+}
 
 const updateCustomMap = async (req: NextApiRequest, res: NextApiResponse) => {
   const userId = await getUserId(req, res)
@@ -56,6 +64,8 @@ const updateCustomMap = async (req: NextApiRequest, res: NextApiResponse) => {
     updatedMap['isPublished'] = isPublished
   }
 
+  updatedMap.lastUpdatedAt = new Date()
+
   const result = await collections.maps?.findOneAndUpdate({ _id: new ObjectId(mapId) }, { $set: updatedMap })
 
   if (!result) {
@@ -64,6 +74,10 @@ const updateCustomMap = async (req: NextApiRequest, res: NextApiResponse) => {
 
   // HALP -> Really shouldn't be deleting locations and then inserting new
   if (locations) {
+    if (locations.length > MAX_ALLOWED_CUSTOM_LOCATIONS) {
+      return throwError(res, 400, `The max locations allowed is ${formatLargeNumber(MAX_ALLOWED_CUSTOM_LOCATIONS)}`)
+    }
+
     // Removes old locations
     const removeResult = await collections.userLocations?.deleteMany({ mapId: new ObjectId(mapId) })
